requirements.txt
numpy
sklearn
scipy
Pillow
matplotlib